pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "growthbook-openfeature-provider"
version = "0.1.0"
description = "GrowthBook provider for OpenFeature"
readme = "README.md"
authors = [{ name = "GrowthBook Maintainers", email = "hello@growthbook.io" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "openfeature-sdk>=0.4.0",
    "growthbook @ git+https://github.com/growthbook/growthbook-python.git@feat/multi-context",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/growthbook/growthbook-openfeature-provider"
Issues = "https://github.com/growthbook/growthbook-openfeature-provider/issues"

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3 